backend/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
sqlalchemy
passlib[bcrypt]
python-jose
email-validator
